pyproject.toml
[tool.poetry]
name = "mc-hackathon-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
google-api-python-client = "^2.149.0"
google-auth = "^2.35.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
flask-sqlalchemy = "^3.1.1"
flask-login = "^0.6.3"
pymysql = "^1.1.1"
cryptography = "^43.0.3"
jwt = "^1.3.1"
pyjwt = "^2.9.0"
google-generativeai = "^0.8.3"
vertexai = "^1.70.0"
faker = "^30.6.0"
pillow = "^11.0.0"
flask-cors = "^5.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
